feat: result画面から購入二飛べるようにしました
