+++ natto/View/Select/SelectViewController.xib
@@ -1,0 +1,80 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="17156" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina6_1" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="17125"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="System colors in document resources" minToolsVersion="11.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="SelectViewController" customModule="natto" customModuleProvider="target">
+            <connections>
+                <outlet property="view" destination="i5M-Pr-FkT" id="sfx-zR-JGt"/>
+            </connections>
+        </placeholder>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <view clearsContextBeforeDrawing="NO" contentMode="scaleToFill" id="i5M-Pr-FkT">
+            <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
+            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+            <subviews>
+                <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="xqx-cN-boM">
+                    <rect key="frame" x="352" y="784" width="46" height="46"/>
+                    <color key="backgroundColor" red="0.12941176469999999" green="0.12941176469999999" blue="1" alpha="1" colorSpace="calibratedRGB"/>
+                    <constraints>
+                        <constraint firstAttribute="height" constant="46" id="Rf1-lP-Iud"/>
+                    </constraints>
+                    <state key="normal" title="Button"/>
+                    <connections>
+                        <action selector="dismissAction:" destination="-1" eventType="touchUpInside" id="8f1-Dc-W5H"/>
+                    </connections>
+                </button>
+                <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="I9J-NR-svX">
+                    <rect key="frame" x="264" y="700" width="80" height="80"/>
+                    <color key="backgroundColor" red="1" green="0.14901960780000001" blue="0.14901960780000001" alpha="1" colorSpace="calibratedRGB"/>
+                    <constraints>
+                        <constraint firstAttribute="width" constant="80" id="RsD-qR-N1H"/>
+                        <constraint firstAttribute="height" constant="80" id="n2S-qr-NvR"/>
+                    </constraints>
+                    <state key="normal" title="Button"/>
+                    <connections>
+                        <action selector="openStorePage:" destination="-1" eventType="touchUpInside" id="2yY-Ac-59v"/>
+                    </connections>
+                </button>
+                <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="7lL-1u-XBk" customClass="ShareButton" customModule="natto" customModuleProvider="target">
+                    <rect key="frame" x="202" y="784" width="40" height="40"/>
+                    <constraints>
+                        <constraint firstAttribute="height" constant="40" id="CCX-Ts-56Y"/>
+                        <constraint firstAttribute="width" constant="40" id="cJi-y8-BHV"/>
+                    </constraints>
+                    <state key="normal" title="Button"/>
+                    <connections>
+                        <action selector="shareAction:" destination="-1" eventType="touchUpInside" id="hQ6-di-UJF"/>
+                    </connections>
+                </button>
+            </subviews>
+            <viewLayoutGuide key="safeArea" id="fnl-2z-Ty3"/>
+            <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+            <constraints>
+                <constraint firstItem="7lL-1u-XBk" firstAttribute="top" secondItem="I9J-NR-svX" secondAttribute="bottom" constant="4" id="7WE-5R-Dv4"/>
+                <constraint firstItem="fnl-2z-Ty3" firstAttribute="trailing" secondItem="xqx-cN-boM" secondAttribute="trailing" constant="16" id="Hm2-Vq-cZI"/>
+                <constraint firstItem="xqx-cN-boM" firstAttribute="top" secondItem="I9J-NR-svX" secondAttribute="bottom" constant="4" id="NL2-1E-MQe"/>
+                <constraint firstItem="I9J-NR-svX" firstAttribute="leading" secondItem="7lL-1u-XBk" secondAttribute="trailing" constant="22" id="Qjl-va-Yhz"/>
+                <constraint firstItem="fnl-2z-Ty3" firstAttribute="trailing" secondItem="I9J-NR-svX" secondAttribute="trailing" constant="70" id="itx-wd-95V"/>
+                <constraint firstItem="fnl-2z-Ty3" firstAttribute="bottom" secondItem="xqx-cN-boM" secondAttribute="bottom" constant="32" id="u85-oz-G05"/>
+            </constraints>
+            <point key="canvasLocation" x="118.84057971014494" y="89.732142857142847"/>
+        </view>
+    </objects>
+    <designables>
+        <designable name="7lL-1u-XBk">
+            <size key="intrinsicContentSize" width="46" height="30"/>
+        </designable>
+    </designables>
+    <resources>
+        <systemColor name="systemBackgroundColor">
+            <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+        </systemColor>
+    </resources>
+</document>
