+++ natto/View/Store/StoreViewController.xib
@@ -45,6 +45,17 @@
                         <action selector="buyAction:" destination="-1" eventType="touchUpInside" id="7r3-SZ-GPq"/>
                     </connections>
                 </button>
+                <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="Qca-hW-6fj" customClass="CloseButton" customModule="natto" customModuleProvider="target">
+                    <rect key="frame" x="368" y="60" width="30" height="30"/>
+                    <constraints>
+                        <constraint firstAttribute="width" constant="30" id="3f2-fh-4S3"/>
+                        <constraint firstAttribute="height" constant="30" id="wDU-Lc-XCw"/>
+                    </constraints>
+                    <state key="normal" title="Button"/>
+                    <connections>
+                        <action selector="dismissAction:" destination="-1" eventType="touchUpInside" id="YVM-PX-xQ1"/>
+                    </connections>
+                </button>
             </subviews>
             <viewLayoutGuide key="safeArea" id="fnl-2z-Ty3"/>
             <color key="backgroundColor" systemColor="systemBackgroundColor"/>
@@ -54,12 +65,19 @@
                 <constraint firstItem="OUc-2P-nLw" firstAttribute="leading" secondItem="fnl-2z-Ty3" secondAttribute="leading" constant="16" id="DTB-5S-vR3"/>
                 <constraint firstItem="fnl-2z-Ty3" firstAttribute="trailing" secondItem="OUc-2P-nLw" secondAttribute="trailing" constant="16" id="PZv-Db-gRK"/>
                 <constraint firstItem="unW-FN-42d" firstAttribute="centerY" secondItem="i5M-Pr-FkT" secondAttribute="centerY" id="gjo-g3-ygZ"/>
+                <constraint firstItem="Qca-hW-6fj" firstAttribute="top" secondItem="fnl-2z-Ty3" secondAttribute="top" constant="16" id="hFo-PN-EL3"/>
                 <constraint firstItem="sGj-VX-6RB" firstAttribute="top" secondItem="OUc-2P-nLw" secondAttribute="bottom" constant="16" id="jwr-yh-rAF"/>
                 <constraint firstItem="sGj-VX-6RB" firstAttribute="centerX" secondItem="OUc-2P-nLw" secondAttribute="centerX" id="qfZ-aM-hlI"/>
+                <constraint firstItem="fnl-2z-Ty3" firstAttribute="trailing" secondItem="Qca-hW-6fj" secondAttribute="trailing" constant="16" id="sq6-rp-zWt"/>
             </constraints>
             <point key="canvasLocation" x="120" y="95"/>
         </view>
     </objects>
+    <designables>
+        <designable name="Qca-hW-6fj">
+            <size key="intrinsicContentSize" width="46" height="30"/>
+        </designable>
+    </designables>
     <resources>
         <image name="negi" width="100" height="100"/>
         <systemColor name="systemBackgroundColor">
